+++ hydromt/gis_utils.py
@@ -5,6 +5,7 @@
 import glob
 import logging
 import os
+from io import IOBase
 from os.path import dirname
 from typing import Optional, Tuple
 
@@ -12,6 +13,7 @@ import geopandas as gpd
 import numpy as np
 import xarray as xr
 from pyflwdir import gis_utils as gis
+from pyogrio import read_info
 from pyproj import CRS
 from pyproj.transformer import Transformer
 from rasterio.transform import Affine
@@ -64,6 +66,9 @@ GDAL_DRIVER_CODE_MAP = {
     "xyz": "XYZ",
 }
 GDAL_EXT_CODE_MAP = {v: k for k, v in GDAL_DRIVER_CODE_MAP.items()}
+
+GPD_TYPES = gpd.GeoDataFrame | gpd.GeoSeries
+GEOM_TYPES = GPD_TYPES | BaseGeometry
 
 ## GEOM functions
 
@@ -553,3 +558,48 @@ def to_geographic_bbox(bbox, source_crs):
         bbox = Transformer.from_crs(source_crs, target_crs).transform_bounds(*bbox)
 
     return bbox
+
+
+def bbox_from_file_and_filters(
+    fn: IOBase,
+    bbox: GEOM_TYPES | None = None,
+    mask: GEOM_TYPES | None = None,
+    crs: CRS | None = None,
+) -> Tuple[float, float, float, float] | None:
+    """Create a bbox from the file metadata and filter options.
+
+    Pyogrio does not accept a mask, and requires a bbox in the same CRS as the data.
+    This function takes the possible bbox filter, mask filter and crs of the input data
+    and returns a bbox in the same crs as the data based on the input filters.
+    As pyogrio currently does not support filtering using a mask, the mask is converted
+    to a bbox and the bbox is returned so that the data has some geospatial filtering.
+
+    Parameters
+    ----------
+    fn: IOBase,
+        opened file
+    bbox: GeoDataFrame | GeoSeries | BaseGeometry
+        bounding box to filter the data while reading
+    mask: GeoDataFrame | GeoSeries | BaseGeometry
+        mask to filter the data while reading
+    crs: pyproj.CRS
+        coordinate reference system of the bounding box or geometry. If already set,
+        this argument is ignored.
+    """
+    if bbox is not None and mask is not None:
+        raise ValueError(
+            "Both 'bbox' and 'mask' are provided. Please provide only one."
+        )
+    if bbox is None and mask is None:
+        return None
+    source_crs = CRS(read_info(fn).get("crs", "EPSG:4326"))  # assume WGS84
+
+    if mask is not None:
+        bbox = mask
+
+    # convert bbox to geom with input crs (assume WGS84 if not provided)
+    crs = crs if crs is not None else CRS.from_user_input(4326)
+    if issubclass(type(bbox), BaseGeometry):
+        bbox = gpd.GeoSeries(bbox, crs=crs)
+    bbox = bbox if bbox.crs is not None else bbox.set_crs(crs)
+    return tuple(bbox.to_crs(source_crs).total_bounds)