+++ tests/test_gis_utils.py
@@ -1,12 +1,16 @@
 """Test for hydromt.gu submodule."""
 
 import os
+from pathlib import Path
 
+import geopandas as gpd
 import numpy as np
 import pytest
 import xarray as xr
 from affine import Affine
+from pyproj import CRS
 from rasterio.transform import from_origin
+from shapely import Polygon, box
 
 from hydromt import gis_utils as gu
 from hydromt.io import open_raster
@@ -127,3 +131,86 @@ def test_create_vrt(tmpdir, rioda_large):
         gu.create_vrt(vrt_fn)
     with pytest.raises(IOError, match="No files found at "):
         gu.create_vrt(vrt_fn, files_path=os.path.join(path, "dummy_xyz", "*.abc"))
+
+
+class TestBBoxFromFileAndFilters:
+    @pytest.fixture(scope="class")
+    def vector_data_with_crs(self, geodf: gpd.GeoDataFrame, tmp_dir: Path) -> Path:
+        example_data = geodf.set_crs(crs=CRS.from_user_input(4326))
+        example_data.to_crs(crs=CRS.from_user_input(3857), inplace=True)
+        path = tmp_dir / "test.fgb"
+        example_data.to_file(path, engine="pyogrio")
+        return path
+
+    @pytest.fixture(scope="class")
+    def vector_data_without_crs(self, geodf: gpd.GeoDataFrame, tmp_dir: Path) -> Path:
+        path = tmp_dir / "test.geojson"
+        geodf.to_file(path, engine="pyogrio")
+        return path
+
+    @pytest.fixture(scope="class")
+    def gdf_mask_without_crs(self, world: gpd.GeoDataFrame) -> gpd.GeoDataFrame:
+        return world[world["name"] == "Chile"]
+
+    @pytest.fixture(scope="class")
+    def gdf_bbox_with_crs(
+        self, gdf_mask_without_crs: gpd.GeoDataFrame
+    ) -> gpd.GeoDataFrame:
+        return gdf_mask_without_crs.set_crs(CRS.from_user_input(4326))
+
+    @pytest.fixture(scope="class")
+    def shapely_bbox(self, gdf_mask_without_crs: gpd.GeoDataFrame) -> Polygon:
+        return box(*list(gdf_mask_without_crs.total_bounds))
+
+    def test_gdf_bbox_crs_source_crs(
+        self, gdf_bbox_with_crs: gpd.GeoDataFrame, vector_data_with_crs: Path
+    ):
+        bbox = gu.bbox_from_file_and_filters(
+            vector_data_with_crs, bbox=gdf_bbox_with_crs
+        )
+        # assert converted to CRS of source data EPSG:3857
+        assert all(map(lambda x: abs(x) > 180, bbox))
+
+    def test_gdf_mask_no_crs_source_crs(
+        self, gdf_mask_without_crs: gpd.GeoDataFrame, vector_data_with_crs: Path
+    ):
+        bbox = gu.bbox_from_file_and_filters(
+            vector_data_with_crs, bbox=gdf_mask_without_crs
+        )
+        # assert converted to CRS of source data EPSG:3857
+        assert all(map(lambda x: abs(x) > 180, bbox))
+
+    def test_gdf_mask_crs_source_no_crs(
+        self, gdf_mask_without_crs: gpd.GeoDataFrame, vector_data_without_crs: Path
+    ):
+        bbox = gu.bbox_from_file_and_filters(
+            vector_data_without_crs, bbox=gdf_mask_without_crs
+        )
+        assert all(map(lambda x: abs(x) < 180, bbox))
+
+    def test_gdf_mask_no_crs_source_no_crs(
+        self, gdf_mask_without_crs: gpd.GeoDataFrame, vector_data_without_crs: Path
+    ):
+        bbox = gu.bbox_from_file_and_filters(
+            vector_data_without_crs, bbox=gdf_mask_without_crs, crs=4326
+        )
+        assert all(map(lambda x: abs(x) < 180, bbox))
+
+    def test_shapely_input(self, shapely_bbox: Polygon, vector_data_with_crs: Path):
+        bbox = gu.bbox_from_file_and_filters(vector_data_with_crs, bbox=shapely_bbox)
+        assert all(map(lambda x: abs(x) > 180, bbox))
+
+    def test_does_not_filter(self, vector_data_with_crs: Path):
+        bbox = gu.bbox_from_file_and_filters(vector_data_with_crs)
+        assert bbox is None
+
+    def test_raises_valueerror(
+        self, vector_data_with_crs: Path, gdf_bbox_with_crs: gpd.GeoDataFrame
+    ):
+        with pytest.raises(
+            ValueError,
+            match="Both 'bbox' and 'mask' are provided. Please provide only one.",
+        ):
+            gu.bbox_from_file_and_filters(
+                vector_data_with_crs, bbox=gdf_bbox_with_crs, mask=gdf_bbox_with_crs
+            )