+++ docs/changelog.rst
@@ -81,6 +81,7 @@ Model
 Vector
 ^^^^^^
 - ``vector.GeoDataset.from_gdf`` can use the gdf columns as data_vars instead of external xarray. (PR #412)
+- ``io.open_vector`` now can use `pyogrio` if reading from a non-tabular dataset (PR #583)
 
 Fixed
 -----